update DB dumps. intention is to eliminate database_dump.sql, in favor of database_wipe_dump.sql.
database_wipe_dump.sql is generate in phpMyAdmin with the following settings:
(most of these are defaults, the main option switched is checking "Add DROP TABLE...", and turning off "IF NOT EXISTS". this is to achieve the wipe. future goal: only wipe ci_bp_meta, and create a delta function to handle alters/upgrades)
+ All tables
+ Options
    - Comments: check
    - Structure: check
         - Add DROP TABLE / VIEW / PROCEDURE / FUNCTION / EVENT: check
         - Add AUTO_INCREMENT value: check
         - Enclose table and field names with backquotes: check
    - Data: check
        - Complete inserts: check
        - Extended inserts: check
        - Use hexadecimal for BLOB: check
        - Export type: INSERT: check