The HMVC upgrade. we moved a whole bunch of files (outside of git, not a big deal really) and added the HMVC library (project URL: https://bitbucket.org/wiredesignz/codeigniter-modular-extensions-hmvc/)
In terms of a broadwa description:
Blog has been fully moved to modules and functions as it should.
Auth controller & views has been moved into modules, but there's still a matter of the views/email directory.
also, we added a crapload of index.html files, in accordance with CI protocol to block access to directory view. this really should be managed via htaccess, but i suppose this provides a more fool-proof environment from the get-go.